% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip6_best_grid}
\alias{cmip6_best_grid}
\title{Get the "best" grid}
\usage{
cmip6_best_grid(catalogue)
}
\arguments{
\item{catalogue}{A catalogue returned by \code{\link[=cmip6]{cmip6()}}}
}
\description{
This selects the best grid based on preferring regridded
data to the preferred target grid ("gr") and then any
other regridding ("gr1/gr2") and fall back to native grid
"gn" if there's no other choice.
}
