% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip6_latest}
\alias{cmip6_latest}
\title{Get latest version}
\usage{
cmip6_latest(catalogue)
}
\arguments{
\item{catalogue}{A catalogue returned by \code{\link[=cmip6]{cmip6()}}}
}
\description{
Returns the latest version of each CMIP6 file
}
