% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadiera5-package.R
\docType{package}
\name{gadiera5-package}
\alias{gadiera5}
\alias{gadiera5-package}
\title{gadiera5: Retrieves ERA5 files from gadi replication}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Elio Campitelli \email{eliocampitelli@gmail.com} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
