% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{latest_version}
\alias{latest_version}
\title{Get latest version}
\usage{
latest_version(catalogue)
}
\arguments{
\item{catalogue}{A catalogue returned by \code{\link[=cmip6]{cmip6()}}}
}
\description{
Returns the latest version of each CMIP6 file
}
