% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip6_latest}
\alias{cmip6_latest}
\alias{cmip6_best_grid}
\title{Quick filters}
\usage{
cmip6_latest(catalogue)

cmip6_best_grid(catalogue)
}
\arguments{
\item{catalogue}{A catalogue returned by \code{\link[=cmip6]{cmip6()}}}
}
\description{
Quick filters
}
\details{
\code{cmip6_latest()} reutrns the latest version of each file.
\code{cmip6_best_grid()} returns the "best" grid based on
preferring data regridded to the preferred target grid ("gr")
and then any other regridding ("gr1/gr2") and fall back to native grid
"gn" if there's no other choice.

These functions ignore the \code{grid_label} and \code{version} columns respectively.
If there are multiple versions with different
grids each, \code{cmip6_latest()} will return the latest version
irrespective of grid and \code{cmip6_best_grid()} will return the best grid
irrespective of version.
This means that the this functions are \emph{not} conmutative.
}
